+++ differt2d/abc.py
@@ -57,16 +57,24 @@ class Plottable(ABC):
     @eqx.filter_jit
     @jaxtyped(typechecker=typechecker)
     def grid(
-        self, n: int = 50
-    ) -> tuple[Float[Array, "{n} {n}"], Float[Array, "{n} {n}"]]:
+        self,
+        m: int = 50,
+        n: Optional[int] = None,
+    ) -> tuple[Float[Array, "n_or_m {m}"], Float[Array, "n_or_m {m}"]]:
         """
         Returns a (mesh) grid that overlays the current object.
 
-        :param n: The number of sample along one axis.
+        :param m: The number of sample along x dimension.
+        :param n: The number of sample along y dimension,
+            defaults to ``m`` is left unspecified.
         :return: A tuple of (X, Y) coordinates.
         """
         bounding_box = self.bounding_box()
-        x = jnp.linspace(bounding_box[0, 0], bounding_box[1, 0], n)
+
+        if n is None:
+            n = m
+
+        x = jnp.linspace(bounding_box[0, 0], bounding_box[1, 0], m)
         y = jnp.linspace(bounding_box[0, 1], bounding_box[1, 1], n)
 
         X, Y = jnp.meshgrid(x, y)